# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API for managing translations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List translations with optional filters",
                "operationId": "0436b18d23fb2409c24e5792abb2baf1",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Filter by tags",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "web"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter by key",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "greeting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of translations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "locale": {
                                                "type": "string",
                                                "example": "en"
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "greeting"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Hello!"
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "web"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "51818ae686a9f08bb44903ebb1eb616c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "welcome"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Welcome!"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "web"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "locale": {
                                            "type": "string",
                                            "example": "en"
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "welcome"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Welcome!"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "web"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get a specific translation",
                "operationId": "a9fc35fcc376efc589f0511c32ff165c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the translation to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "locale": {
                                            "type": "string",
                                            "example": "en"
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "greeting"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Hello!"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "web"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update an existing translation",
                "operationId": "fb5c450fb84f94732d60db9d5c705020",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the translation to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "greeting"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Hello, world!"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "mobile"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "locale": {
                                            "type": "string",
                                            "example": "en"
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "greeting"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Hello, world!"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "mobile"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export all translations as JSON",
                "operationId": "0400869a94a5f8b83854c6a50bdee05e",
                "responses": {
                    "200": {
                        "description": "Exported translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "locale": {
                                                "type": "string",
                                                "example": "en"
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "greeting"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Hello!"
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "web"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Translations",
            "description": "Translations"
        }
    ]
}